+++ lib/src/tabs_area.dart
@@ -1,14 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-import 'package:tabbed_view/src/flow_layout.dart';
-import 'package:tabbed_view/src/tab_button.dart';
-import 'package:tabbed_view/src/tab_button_widget.dart';
-import 'package:tabbed_view/src/tab_data.dart';
+import 'package:tabbed_view/src/internal/tabbed_view_provider.dart';
+import 'package:tabbed_view/src/internal/tabs_area/hidden_tabs.dart';
+import 'package:tabbed_view/src/internal/tabs_area/tabs_area_corner.dart';
 import 'package:tabbed_view/src/tab_status.dart';
 import 'package:tabbed_view/src/tab_widget.dart';
 import 'package:tabbed_view/src/tabbed_view_controller.dart';
-import 'package:tabbed_view/src/tabbed_view_data.dart';
-import 'package:tabbed_view/src/tabbed_view_menu_item.dart';
 import 'package:tabbed_view/src/tabs_area_layout.dart';
 import 'package:tabbed_view/src/theme/tabbed_view_theme_data.dart';
 import 'package:tabbed_view/src/theme/tabs_area_theme_data.dart';
@@ -16,9 +12,9 @@ import 'package:tabbed_view/src/theme/theme_widget.dart';
 
 /// Widget for the tabs and buttons.
 class TabsArea extends StatefulWidget {
-  const TabsArea({required this.data});
+  const TabsArea({required this.provider});
 
-  final TabbedViewData data;
+  final TabbedViewProvider provider;
 
   @override
   State<StatefulWidget> createState() => _TabsAreaState();
@@ -28,27 +24,32 @@ class TabsArea extends StatefulWidget {
 class _TabsAreaState extends State<TabsArea> {
   int? _highlightedIndex;
 
-  final HiddenTabs hiddenTabs = HiddenTabs();
+  final HiddenTabs _hiddenTabs = HiddenTabs();
 
   @override
   Widget build(BuildContext context) {
-    TabbedViewController controller = widget.data.controller;
+    TabbedViewController controller = widget.provider.controller;
     TabbedViewThemeData theme = TabbedViewTheme.of(context);
     TabsAreaThemeData tabsAreaTheme = theme.tabsArea;
     List<Widget> children = [];
     for (int index = 0; index < controller.tabs.length; index++) {
       TabStatus status = _getStatusFor(index);
       children.add(TabWidget(
+          key: controller.tabs[index].uniqueKey,
           index: index,
           status: status,
-          data: widget.data,
-          updateHighlightedIndex: _updateHighlightedIndex));
+          provider: widget.provider,
+          updateHighlightedIndex: _updateHighlightedIndex,
+          onClose: _onTabClose));
     }
+
+    children.add(
+        TabsAreaCorner(provider: widget.provider, hiddenTabs: _hiddenTabs));
+
     Widget tabsAreaLayout = TabsAreaLayout(
         children: children,
-        buttonsAreaBuilder: _buttonsAreaBuilder,
         theme: theme,
-        hiddenTabs: hiddenTabs,
+        hiddenTabs: _hiddenTabs,
         selectedTabIndex: controller.selectedIndex);
     tabsAreaLayout = ClipRect(child: tabsAreaLayout);
 
@@ -60,89 +61,9 @@ class _TabsAreaState extends State<TabsArea> {
     return Container(child: tabsAreaLayout, decoration: decoration);
   }
 
-  /// Area for buttons like the hidden tabs menu button.
-  ///
-  /// Even if there are no visible buttons, an empty container must be created.
-  Widget _buttonsAreaBuilder(BuildContext context) {
-    TabbedViewThemeData theme = TabbedViewTheme.of(context);
-    TabsAreaThemeData tabsAreaTheme = theme.tabsArea;
-    Widget buttonsArea;
-
-    List<TabButton> buttons = [];
-    if (widget.data.tabsAreaButtonsBuilder != null) {
-      buttons = widget.data.tabsAreaButtonsBuilder!(
-          context, widget.data.controller.tabs.length);
-    }
-
-    if (hiddenTabs.hasHiddenTabs) {
-      TabButton hiddenTabsMenuButton = TabButton(
-          icon: tabsAreaTheme.menuIcon, menuBuilder: _hiddenTabsMenuBuilder);
-      buttons.insert(0, hiddenTabsMenuButton);
-    }
-
-    if (buttons.isNotEmpty) {
-      List<Widget> children = [];
-      for (int i = 0; i < buttons.length; i++) {
-        EdgeInsets? padding;
-        if (i > 0 && tabsAreaTheme.buttonsGap > 0) {
-          padding = EdgeInsets.only(left: tabsAreaTheme.buttonsGap);
-        }
-        TabButton tabButton = buttons[i];
-        children.add(Container(
-            child: TabButtonWidget(
-                data: widget.data,
-                button: tabButton,
-                enabled: true,
-                normalColor: tabsAreaTheme.normalButtonColor,
-                hoverColor: tabsAreaTheme.hoverButtonColor,
-                disabledColor: tabsAreaTheme.disabledButtonColor,
-                normalBackground: tabsAreaTheme.normalButtonBackground,
-                hoverBackground: tabsAreaTheme.hoverButtonBackground,
-                disabledBackground: tabsAreaTheme.disabledButtonBackground,
-                iconSize: tabButton.iconSize != null
-                    ? tabButton.iconSize!
-                    : tabsAreaTheme.buttonIconSize,
-                themePadding: tabsAreaTheme.buttonPadding),
-            padding: padding));
-      }
-
-      buttonsArea = FlowLayout(children: children, firstChildFlex: false);
-
-      EdgeInsetsGeometry? margin;
-      if (tabsAreaTheme.buttonsOffset > 0) {
-        margin = EdgeInsets.only(left: tabsAreaTheme.buttonsOffset);
-      }
-      if (tabsAreaTheme.buttonsAreaDecoration != null ||
-          tabsAreaTheme.buttonsAreaPadding != null ||
-          margin != null) {
-        buttonsArea = Container(
-            child: buttonsArea,
-            decoration: tabsAreaTheme.buttonsAreaDecoration,
-            padding: tabsAreaTheme.buttonsAreaPadding,
-            margin: margin);
-      }
-    } else {
-      buttonsArea = SizedBox(width: 0);
-    }
-    return buttonsArea;
-  }
-
-  /// Builder for hidden tabs menu.
-  List<TabbedViewMenuItem> _hiddenTabsMenuBuilder(BuildContext context) {
-    List<TabbedViewMenuItem> list = [];
-    hiddenTabs.indexes.sort();
-    for (int index in hiddenTabs.indexes) {
-      TabData tab = widget.data.controller.tabs[index];
-      list.add(TabbedViewMenuItem(
-          text: tab.text,
-          onSelection: () => widget.data.controller.selectedIndex = index));
-    }
-    return list;
-  }
-
   /// Gets the status of the tab for a given index.
   TabStatus _getStatusFor(int tabIndex) {
-    TabbedViewController controller = widget.data.controller;
+    TabbedViewController controller = widget.provider.controller;
     if (controller.tabs.isEmpty || tabIndex >= controller.tabs.length) {
       throw Exception('Invalid tab index: $tabIndex');
     }
@@ -162,5 +83,11 @@ class _TabsAreaState extends State<TabsArea> {
         _highlightedIndex = tabIndex;
       });
     }
+  }
+
+  void _onTabClose() {
+    setState(() {
+      _highlightedIndex = null;
+    });
   }
 }