+++ CHANGELOG.md
@@ -1,3 +1,79 @@
+## 1.18.0
+
+* Highlighting the tab's drop position.
+* Allow dragging to reorder tabs to the last position.
+* Allow dragging tabs between different `TabbedView` instances.
+* Bugfix
+  * Incorrect state for the index of the highlighted tab after being closed.
+
+### Changes
+
+* Change in the signature of `OnDraggableBuild`.
+  * From: `(int tabIndex, TabData tabData)`
+  * To: `(TabbedViewController controller, int tabIndex, TabData tabData)`
+* `Draggable` will always be `DraggableData` type: `Draggable<DraggableData>`
+* `TabsAreaThemeData`
+  * New attribute: `dropColor`.
+
+## 1.17.0
+
+* Tab reordering
+* `TabbedViewController`
+  * New methods
+    * `setTabs`
+    * `reorderTab`
+    * `selectedTab`
+  * New attribute: `reorderEnable`
+* New callback: `OnReorder` 
+* `TabThemeData`
+  * New attributes
+    * `draggingDecoration`
+    * `draggingOpacity`
+* `TabData`
+  * New attribute: `draggable`
+* New class: `DraggableConfig`
+* New typedef: `OnDraggableBuild`
+* `TabbedView`
+  * The `draggableTabBuilder` has been replaced by `onDraggableBuild`
+    * Automatic creation of a `Draggable<TabData>`
+* `TabData`
+  * The `uniqueKey` attribute has been renamed to `key`.
+* Minimum sdk version required: 2.19.0
+
+### Migrating custom drag feedback
+
+From:
+```dart
+    TabbedView tabbedView = TabbedView(
+        controller: _controller,
+        draggableTabBuilder: (int tabIndex, TabData tab, Widget tabWidget) {
+          return Draggable<String>(
+              child: tabWidget,
+              feedback: Material(
+                  child: Container(
+                      child: Text(tab.text),
+                      padding: EdgeInsets.all(4),
+                      decoration: BoxDecoration(border: Border.all()))),
+              data: tab.text,
+              dragAnchorStrategy: (Draggable<Object> draggable,
+                  BuildContext context, Offset position) {
+                return Offset.zero;
+              });
+        });
+```
+To:
+```dart
+    TabbedView tabbedView = TabbedView(
+        controller: _controller,
+        onDraggableBuild: (int tabIndex, TabData tabData) {
+          return DraggableConfig(
+              feedback: Container(
+                  child: Text(tabData.text),
+                  padding: EdgeInsets.all(4),
+                  decoration: BoxDecoration(border: Border.all())));
+        });
+```
+
 ## 1.16.0+1
 
 * Removing the use of deprecated constructor.